+++ jsonpickle/pickler.py
@@ -7,6 +7,8 @@
 # This software is licensed as described in the file COPYING, which
 # you should have received as part of this distribution.
 
+from itertools import chain
+
 import jsonpickle.util as util
 import jsonpickle.tags as tags
 import jsonpickle.handlers as handlers
@@ -51,10 +53,13 @@ class Pickler(object):
         self._max_depth = max_depth
         ## Maps id(obj) to reference IDs
         self._objs = {}
+        ## Avoids garbage collection
+        self._seen = []
 
     def reset(self):
         self._objs = {}
         self._depth = -1
+        self._seen = []
 
     def _push(self):
         """Steps down one level in the namespace.
@@ -119,7 +124,10 @@ class Pickler(object):
 
     def _flatten(self, obj):
         self._push()
+        return self._pop(self._flatten_obj(obj))
 
+    def _flatten_obj(self, obj):
+        self._seen.append(obj)
         max_reached = self._depth == self._max_depth
 
         if max_reached or (not self.make_refs and id(obj) in self._objs):
@@ -128,14 +136,17 @@ class Pickler(object):
         else:
             flatten_func = self._get_flattener(obj)
 
-        return self._pop(flatten_func(obj))
+        return flatten_func(obj)
+
+    def _list_recurse(self, obj):
+        return [self._flatten(v) for v in obj]
 
     def _get_flattener(self, obj):
 
         if util.is_primitive(obj):
             return lambda obj: obj
 
-        list_recurse = lambda obj: [self._flatten(v) for v in obj]
+        list_recurse = self._list_recurse
 
         if util.is_list(obj):
             if self._mkref(obj):
@@ -186,7 +197,10 @@ class Pickler(object):
         has_class = hasattr(obj, '__class__')
         has_dict = hasattr(obj, '__dict__')
         has_slots = not has_dict and hasattr(obj, '__slots__')
-        has_getstate = has_dict and hasattr(obj, '__getstate__')
+
+        # Support objects with __getstate__(); this ensures that
+        # both __setstate__() and __getstate__() are implemented
+        has_getstate = hasattr(obj, '__getstate__')
         has_getstate_support = has_getstate and hasattr(obj, '__setstate__')
 
         if has_class and not util.is_module(obj):
@@ -207,22 +221,18 @@ class Pickler(object):
             return data
 
         if util.is_dictionary_subclass(obj):
-            return self._flatten_dict_obj(obj, data)
+            self._flatten_dict_obj(obj, data)
+            if has_getstate_support:
+                self._getstate(obj, data)
+            return data
 
         if has_dict:
             # Support objects that subclasses list and set
             if util.is_sequence_subclass(obj):
                 return self._flatten_sequence_obj(obj, data)
 
-            # Support objects with __getstate__(); this ensures that
-            # both __setstate__() and __getstate__() are implemented
             if has_getstate_support:
-                state = self._flatten(obj.__getstate__())
-                if self.unpicklable:
-                    data[tags.STATE] = state
-                else:
-                    data = state
-                return data
+                return self._getstate(obj, data)
 
             # hack for zope persistent objects; this unghostifies the object
             getattr(obj, '_', None)
@@ -256,8 +266,15 @@ class Pickler(object):
     def _flatten_newstyle_with_slots(self, obj, data):
         """Return a json-friendly dict for new-style objects with __slots__.
         """
-        for k in obj.__slots__:
-            self._flatten_key_value_pair(k, getattr(obj, k), data)
+        allslots = [getattr(cls, '__slots__', tuple())
+                        for cls in type(obj).mro()]
+        for k in chain(*allslots):
+            try:
+                value = getattr(obj, k)
+            except AttributeError:
+                # The attribute may have been deleted
+                continue
+            self._flatten_key_value_pair(k, value, data)
         return data
 
     def _flatten_key_value_pair(self, k, v, data):
@@ -288,6 +305,15 @@ class Pickler(object):
         else:
             return value
         return data
+
+    def _getstate(self, obj, data):
+        state = self._flatten_obj(obj.__getstate__())
+        if self.unpicklable:
+            data[tags.STATE] = state
+        else:
+            data = state
+        return data
+
 
 def _mktyperef(obj):
     """Return a typeref dictionary